Started work on the YAML parser.
So as not to disturb the stable upstream, I'll be modifying the grader on the "yaml" branch. It shouldn't take too long to complete, and I'll also add in more features (I kinda have to in order to get the flexible YAML parsing to work). It won't be runnable for some time, which is the crux of the reason of making this new branch.