Made Checks.java resiliant (again) to client failure.
Figured out that the run-mdb-server.sh script was blocking due to the fact that the mdb-lookup-server still had potential output (it needed to be redirected to /dev/null).
Also, removed the stupid synchronization from Checks.java, as it's utility is limited to the case where a thread spuriously dies (which makes no sense).